+++ py_bootstrap/templates/application/__metadata__.py
@@ -1,18 +1,16 @@
-DESCRIPTION = (
-    "Provides bootstrapping for Python applications based on tox tool."
-)
+DESCRIPTION = "Generates a skeleton of Python application based on tox tool."
 
-CONTEXT = {
-    "name": None,
-    "dir_name": None,
-    "title": None,
-    "description": None,
-    "repo": None,
+CONTEXT: dict[str, str] = {
+    "name": "",
+    "dir_name": "",
+    "title": "",
+    "description": "",
+    "repo": "",
     "python_major": "3",
     "python_minor": "13",
 }
 
-CONTEXT_REQUIRED_KEYS = [
+CONTEXT_REQUIRED_KEYS: list[str] = [
     "name",
     "description",
 ]
@@ -29,10 +27,10 @@ def prepare_context(context: dict[str, str]):
     import re
     from datetime import datetime
 
+    context["empty"] = ""
+
     name = context["name"]
     name_parts: list[str] = re.split("[-_ ]", name.strip())
-
-    context["empty"] = ""
 
     if not context.get("dir_name"):
         val = "_".join(name_parts)